Fix duplicate key error collection changes
PushPull saves changes sent by the client then returns changes that the
client did not receive in server. After processing PushPull, the server
saves changes, the document, the client in the DB in order. After saving
the changes, the final seq of the changes is saved in the document. We
didn't use transactions here. Because we intended to overwrite changes
even if the server failed to update seq in the document after saving the
changes.

Relevant ticket: https://github.com/hackerwins/yorkie/issues/8