Add increase operation (#71)
* Add increase operation

I add 'number_proxy.go' and 'increase.go' for increase operation
on 'Integer' and 'Long' types.

* Improving increase operation

Improved to be able to operate with various types of operands,
and improved test code accordingly. And removed unnecessary
'prevCreatedAt' attribute.
In addition, by adding 'GetDouble', the increase operation
for the float64 type has been improved.

* Add increase operation in proto

We need to implement a communication layer between
the client and the server for increase operation.
So I add 'increase test' in 'client_test.go'
and add 'increase operation' in 'yorkie.proto'.

* Improve the increase method of Primitive

Fixed an error in the numeric type check logic. And, when a primitive
of a different number type is entered as an operand of the Increase
method, it is improved to operate normally.