Fix NewCounter method
When integers are used by default in golang, they have an int type.
So, in NewCounter, numbers greater than max int32 are not classified as
int64, but are used as int. And this causes a problem because it is
converted to an incorrect value when calling the Bytes() method before
the server request.