FIx NumberProxy to Counter (#82)
* Change NumberProxy to Counter

Improved so that 'Increase operation' can be performed with a separate
type of Counter like 'Automerge'

* Fix RPCPort of testhelper

During the test, an error occurred due to overlapping with MetricsPort.

* Improve Increase method of counter_proxy

IsNumericType method is used to check the types allowed
for increase operation.

* Fix NewCounter method

When integers are used by default in golang, they have an int type.
So, in NewCounter, numbers greater than max int32 are not classified as
int64, but are used as int. And this causes a problem because it is
converted to an incorrect value when calling the Bytes() method before
the server request.

* Add concurrent increase operation of counter test

* Add new counter test in counter_test

* Fix to test concurrency of increase operation, not set operation

* Fix Increase method of counter

If the Counter type is an integer type and is greater than MaxInt32 or
less than MinInt32, the counter value will change when calling
the Bytes() method.

* Change operand type of increase operation

Change operand type to Primitive type. Because it is awkward to use
Counter as an operand.

* Update comment of Increase method

Counter's value type can be changed Integer to Long.
Added content for this.

Co-authored-by: hackerwins <susukang98@gmail.com>