Connect adjacent insNext and insPrev when running purge (#123)
Connect adjacent insNext and insPrev when running purge.

In #104, GC was also introduced for text nodes marked with tombstones. 
However, in release(), the link between adjacent insPrev and insNext is 
missing, so this commit fixes it.

`insPrev` and `insNext` are used to remember other nodes connected to 
the insert. For example, when abc is divided into a, b, c, it looks like
 this: [abc] divided to [a]<->[b]<->[c]. Even if we completely delete b,
 I think it should keep the relationship of [a] <-> [c].

And disconnecting `node.prev` and `node.next` is missing when purging a 
node from RGATreeList. So this commit added it.