Isolate clients and documents for each project (#319)
To isolate clients and documents for each project this commit added
`ProjectID` to `DocInfo` and `ClientInfo`. The logic to check whether
the Document and Client belong to the Project is in the database layer.

And keys for clients and documents are unique for each project rather
than a system global. This is to allow users to manage keys by their
projects.

It is not compatible with the existing MongoDB and the migration script
was not written as it was pre-deployment.

This commit also cleans up codes:

- Rename `db` package to `database`
- Extract Paging
- Move project logic to `projects` package
- Remove unused codes