Splay tree Deletion logic improvement
There was two problem in func deleteNodes. 1. repeat splay for
each node to be deleted resulting in numerous rotations. 2. not
actually delete node, zero-weight node left and affect.
improve the deletion logic : to skip splay for a node in between,
and to isolate in a subtree then cut off the subtree.