Fix incorrect index for nodes newly created then concurrently removed (#364)
* Design docs about Splay Tree range deletion

This commit creates a design document about Splay Tree range deletion.
And to increase visibility, the direction of the operation is reversed.
It reduces the code of the function but does not change the external behavior.

* Fix index bug for nodes newly created then removed

Nodes newly created and then concurrently removed in the range to be deleted
had incorrect weight since they were not included in either nodesToKeep
or nodesToDelete. This problem causes incorrect indexing, and this commit
fixes the missing weight updates.