Add SearchDocuments API (#363)
This commit deletes sorting in id order because it would be better to
keep logic simple and apply search engines if various sorting criteria are needed.
Cursor-based pagination is not possible while returning the results sorted by key.
Instead of offset-based pagination, it returns up to given size search results.