Introduce buf (#382)
This commit introduces buf in order to write proto files while following
best practices.

The linter requires that the zero value for all enums should be suffixed
with _UNSPECIFIED. However, it is difficult to migrate the operations
already stored in the DB, so we ignore the rule for now.

We are also considering introducing an IDL repository or Buf Schema
Registry to manage proto files in a single location. Even now, this
repository is SSOT for proto files, but we are using it by manually
copying it from another repository(e.g. yorkie-js-sdk, yorkie-house).

Co-authored-by: Miller <miller@daangn.com>