Extend PushPull to support sync mode by adding push-only flag (#500)
When push-only is set to true, the Server will respond with an empty
operations unless there have been actual changes to the data. In cases
where the push-pull response contains changes that have already been
made on the client side, empty changes will be sent to prevent
redoing those operations.

Due to these modifications, deep copying of changeInfo is required.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>