Remove skip storing MinSyncedTicket when the ticket is initial (#655)
Previously, in order to enhance GC efficiency, when a client initially
attaches a document that has any changes, the process of storing
MinSyncedTicket was skipped. However, this approach is flawed because
the server stores the ticket based on the checkpoint of requests for
handling scenarios where the client's response may be lost.

So this commit removes the optimization for now.

This commit also revises the logs of the PushPull process to be more
simple.