Modify converter to allow setting values for object and array (#687)
PR#678 triggers the regression error, so I made corrections and
re-submitted. I've added handling for cases where values are not set
when converting to PbJSONElementSimple. In the bytesToObject function,
It was previously creating a new object with initialTicket when the
bytes were nil. However, this logic became inappropriate after
separating it from bytesToSnapshot, so I removed it.