Implement splitLevel of Tree.Edit (#705)
When a user positions the cursor within a paragraph and hits the enter
key, the paragraph is divided into two separate paragraphs:

- Original: `<p>a|b</p>`
- Result: `<p>a</p><p>b</p>`

In this commit, `splitLevel` is introduced to support this scenario.

```ts
doc.update((root) => {
  root.tree.edit([0,0], {type:'p',children:[{type:'text', value: 'ab'}]});
  root.tree.edit([2,2], undefined, 1); // split the paragraph
});
```

This commit does not cover all test cases that may arise due to the
split operation. This is because we need to verify the interface
first. we need to add test cases in the simultaneous editing situation
to verify the algorithm later.