Handle concurrent Tree.RemoveStyle (#883)
This commit addresses an issue where if a node has an attribute with
the same key as the RemoveStyle is simultaneously inserted by
Tree.Edit, the attribute of the concurrently inserted Node with the
same key gets deleted. To resolve this issue, this commit adds
filtering logic in RemoveStyle to prevent attribute deletion from
concurrently inserted nodes.