Fix miscalculation of tree size in concurrent editing (#891)
This commit addresses the issue where the updateAncestorsSize method
subtracts the size of a tombstoned node from its ancestors' sizes when
the node is removed. However, the previous logic did not consider
cases where ancestors node are tombstoned. In such cases, the update
should not propagate to the parent's ancestors. This commit introduces
recursive checks on ancestors and stops updating process if a
tombstoned ancestor is encountered.

Furthermore, the PR ensures consistency between the size maintenance
policies of updateDescendantsSize and updateAncestorsSize.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>