Enhance housekeeping to add variety of tasks (#932)
This commit enhances Housekeeping to enable the addition of a variety
of tasks. Previously, Housekeeping was limited to only Client
Deactivation task.

In this update, the actual tasks are implemented in the service
package(e.g. clients), while housekeeping package is only
responsible for scheduling tasks. It is for separation of concerns.