Improve performance for creating crdt.TreeNode (#939)
The time complexity of creating crdt.TreeNode is O(N^2), potentially
causing performance bottlenecks. It's optimized to O(n).

While this may not be a significant issue currently, there is a risk
that as the number of tree nodes in the protobuf increases, operations
will scale quadratically, potentially causing performance bottlenecks.

---------

Co-authored-by: JiHwan Yim <raararaara@gmail.com>
Co-authored-by: Youngteac Hong <susukang98@gmail.com>