Show Server Version in Yorkie CLI (#938)
This commit adds the functionality to display the version of connected
Yorkie server in the version command.

The version information can be viewed by converting it into yaml or
json format using the --output, -o flag. This flag accepts json and
yaml strings and converts them to the appropriate format for display.
In order to solely check the version of the client CLI without
considering the server's version, the --client flag has been included.
The development of this feature was inspired by examining the kubectl.