Add a metric to collect the number of background routines (#963)
This PR adds a metric to collect the number of background routines.

Prometheus provides its own thread-safe Gauge metric, which can be
used to accurately collect the number of concurrent goroutines.

Also, although the only task running in the background so far is
PushPull, this commit implements the AttachGoroutine() method to take
a string called taskType as a parameter, since it seems to be focused
on reusability.