Apply max height splay method
To prevent the tree from becoming skewed, we applied a heuristic cal-
led the 'max height splay' method. When the size of the tree is n,
and its height exceeds 50*(log n), the deepest node is selected and
splayed to the root. To avoid excessive balancing when the tree has
few nodes, the balancing process is skipped if the number of nodes is
1000 or fewer.
Since the 'max height splay' method requires knowledge of the tree's
height, we added a height attribute to the nodes of the splay tree
and included code to check the height during testing.