Detach documents when client is deactivated (#1036)
This commit implements the detachment of documents attached to clients
upon client deactivation. Previously, the status of the document was
only changed to 'detach' within DB.clients.ClientDocInfo.Status,
leading to issues where the deactivated client's presence remained
active and presence change events were not communicated to other
clients.

With this change, on client deactivation, the document's packs.PushPull
method is called directly to execute the complete detachment process.

Moreover, since the Yorkie cluster sharding mechanism assigns specific
documents to specific servers, the server managing client deactivation
may not always be the one handling the document. Therefore, in such
cases, a request will be made to the server currently responsible for
the document to detach it properly. To facilitate this communication
across the cluster, a new ClusterService has been added.

Follow-up tasks that need to be addressed after this PR include:

- Implementing authentication for cluster communication and preventing
  external exposure.
- Eliminating redundant database resource requests between
  clients.Deactivate and server.DetachDocument.
- Improving the p.Clear ChangePack creation logic.
- Introducing a connection pool for clusterClient.

---------

Co-authored-by: raararaara <raararaara@gmail.com>