Fix transaction in UpdateAndFindMinSyncedVersionVector (#1050)
This commit resolves an issue where the PushPull operation does not
function correctly when using the Memory DB.

UpdateAndFindMinSyncedVersionVector was previously set to start in
write mode, which causes it to simply read documents. However, inside
this function, it calls UpdateVersionVector, which attempts to start
in write mode again, resulting in a deadlock scenario where the
transaction cannot be acquired, causing an infinite wait.