Separate correction logic in RegisterGCPair for editing and build modes (#1294)
This commit refactors the RegisterGCPair function by distinguishing
between editing and build modes when applying size correction logic.

In editing mode, size correction is necessary when a node is deleted
and registered in a GCPair.

In build mode (e.g., during document reconstruction from the root),
size correction should not be applied, as it results in incorrect size
calculations.