Replace wildcard index with regular index on attached_docs (#1315)
Wildcard index on the `documents` field of `ClientInfo` was causing the
index size to grow significantly. This is because each document entry in
`documents` contains not only the document key but also additional metadata
such as status, server_seq, and client_seq.

To address this, we added a new field called `attached_docs`, which contains
only the list of attached document keys without additional metadata. This
field is updated alongside `documents` when attaching or detaching a document,
using `$addToSet` and `$pull` operations respectively.