Enhance document attachment with schema validation (#1345)
This commit introduces schema support to validate and manage the structure
of collaborative documents. Schemas can be created and updated via the
dashboard UI, and are stored and validated using the @yorkie-js/schema
package.

A schema can be attached to a document using the SDK before any clients are
attached. If a schema is already attached or the document has active
clients, the attachment is ignored. During local edits, the attached schema
is used to validate changes, and schema violations result in client-side
errors.

Via the AdminService API, schemas can also be updated or detached from
documents. The UpdateDocument API supports different behaviors depending on
whether the `root` and `schemaKey` fields are present or empty.

Currently, only primitive type validations are supported. Additional rule
types will be implemented later in the ruleset generation and validation
logic.