Resolve convergence issues in array operations
This commit addresses convergence issues caused by Array.Move and Array.Set
operations. We introduce a new `movedFrom` field of type *RGATreeListNode
in RGATreeListNode to track the origin of moved elements. New logic is added
to the `add` and `move` operations to support this.

Additionally, the `set` operation is now expressed as a combination of
`InsertAfter(prev, newNode)` and `Remove(node)`, simplifying its behavior and
aligning with existing insertion/removal semantics.