Add feature to filter out stale presences (#1402)
This commit filters stale presences so that only presences of online
clients with established physical streams are displayed. Previously,
stale presences remain visible even when the document is detached, and
the presence clearing mechanism isn't functioning as intended. While
this commit improves the situation by ensuring only relevant presences
are shown, it does not address the root cause of the lingering stale
presences issue.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>